--- src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import Form from './components/form/Form';
-
-class App extends React.Component {
-  render(){
-  return (
-      <Form/>
-  );
-}
-}
-
-export default App;
